# backend/swagger.yaml
openapi: 3.0.0
info:
    description: This is the API documentation for project (Re)SourcesRelationnelles
    version: "1.0.0"
    title: API (Re)SourcesRelationnelles
    contact:
        email: ressources.relationnelles@ressources.fr
    license:
        name: Apache 2.0
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
    - url: /api
      description: API base path
...
paths:
...
    /v1/resource/create:
        post:
            tags:
                - Gestion des ressources
            summary: Création d'une ressource fichier
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - title
                                - category
                                - relationType
                                - file
                            properties:
                                title:
                                    type: string
                                    example: "Guide de santé mentale"
                                category:
                                    type: string
                                    example: "507f1f77bcf86cd799439018"
                                relationType:
                                    type: string
                                    example: "507f1f77bcf86cd799439019"
                                file:
                                    type: string
                                    format: binary
                                    description: "Fichier à uploader (max 500MB)"
            responses:
                "201":
                    description: Ressource créée
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressource créée avec succès"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                $ref: "#/components/schemas/Resource"
...
    /v1/resource/:
        get:
            tags:
                - Gestion des ressources
            summary: Récupération de toutes les ressources (admin)
            description: Récupère toutes les ressources, y compris non publiées (admin/super-admin uniquement)
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Liste des ressources
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressources récupérées avec succès"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                type: array
                                                items:
                                                    $ref: "#/components/schemas/Resource"
    /v1/resource/published/{id}:
        get:
            tags:
                - Gestion des ressources
            summary: Récupération d'une ressource publiée par ID
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                      pattern: "^[0-9a-fA-F]{24}$"
                      example: "507f1f77bcf86cd799439011"
            responses:
                "200":
                    description: Ressource trouvée
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressource trouvée"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                $ref: "#/components/schemas/Resource"
                "404":
                    description: Ressource non trouvée
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /v1/resource/{id}:
        get:
            tags:
                - Gestion des ressources
            summary: Récupération d'une ressource par ID (authentifié)
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                      pattern: "^[0-9a-fA-F]{24}$"
                      example: "507f1f77bcf86cd799439011"
            responses:
                "200":
                    description: Ressource trouvée
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressource trouvée"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                $ref: "#/components/schemas/Resource"

        put:
            tags:
                - Gestion des ressources
            summary: Modification d'une ressource
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                      pattern: "^[0-9a-fA-F]{24}$"
                      example: "507f1f77bcf86cd799439011"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ResourceUpdate"
            responses:
                "200":
                    description: Ressource modifiée
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressource mise à jour avec succès"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                $ref: "#/components/schemas/Resource"

        delete:
            tags:
                - Gestion des ressources
            summary: Suppression d'une ressource
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                      pattern: "^[0-9a-fA-F]{24}$"
                      example: "507f1f77bcf86cd799439011"
            responses:
                "200":
                    description: Ressource supprimée
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SuccessResponse"
    /v1/resource/create-text:
        post:
            tags:
                - Gestion des ressources
            summary: Création d'une ressource texte/HTML
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - title
                                - category
                                - relationType
                                - content
                            properties:
                                title:
                                    type: string
                                    example: "Article sur la gestion du stress"
                                category:
                                    type: string
                                    example: "507f1f77bcf86cd799439018"
                                relationType:
                                    type: string
                                    example: "507f1f77bcf86cd799439019"
                                content:
                                    type: string
                                    description: "Contenu HTML de la ressource"
                                    example: "<h1>Gestion du stress</h1><p>Voici quelques conseils...</p>"
            responses:
                "201":
                    description: Ressource créée
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Ressource créée avec succès"
                                    data:
                                        type: object
                                        properties:
                                            ressource:
                                                $ref: "#/components/schemas/Resource"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        SuccessResponse:
            type: object
            properties:
                message:
                    type: string
                    example: "Opération réussie"
                data:
                    type: object
        ErrorResponse:
            type: object
            properties:
                error:
                    oneOf:
                        - type: object
                          properties:
                              msg:
                                  type: string
                                  example: "Message d'erreur"
                              location:
                                  type: string
                                  example: "body"
                        - type: object
                          properties:
                              errors:
                                  type: array
                                  items:
                                      type: object
                                      properties:
                                          location:
                                              type: string
                                              example: "body"
                                          msg:
                                              type: string
                                              example: "Le champ est requis"
                                          path:
                                              type: string
                                              example: "email"
                                          type:
                                              type: string
                                              example: "field"
                                          value:
                                              type: string
                                              example: "invalid-email"
...
        Resource:
            type: object
            properties:
                _id:
                    type: string
                    example: "507f1f77bcf86cd799439011"
                title:
                    type: string
                    example: "Guide de santé mentale"
                category:
                    type: string
                    example: "507f1f77bcf86cd799439018"
                relationType:
                    type: string
                    example: "507f1f77bcf86cd799439019"
                createdAt:
                    type: string
                    format: date-time
                    example: "2024-01-01T00:00:00Z"
                updatedAt:
                    type: string
                    format: date-time
                    example: "2024-01-01T00:00:00Z"
        ResourceUpdate:
            type: object
            properties:
                title:
                    type: string
                    example: "Guide de santé mentale"
                category:
                    type: string
                    example: "507f1f77bcf86cd799439018"
                relationType:
                    type: string
                    example: "507f1f77bcf86cd799439019"
...